[DF-520] Changed uint8 to uint32 due to uint8 not added to VariantType

+++ Sources/Internal/Render/Highlevel/Light.h
@@ -97,7 +97,7 @@ protected:
     Vector3 direction;
     
 //    eType type; //TODO: waiting for enums at introspection
-    uint8 type;
+    uint32 type;
     
     Color ambientColor;
     Color diffuseColor;
